+++ tests/testthat/helper.R
@@ -1,3 +1,12 @@
+skip_without_gcs_credentials <- function() {
+  # -# Sys.setenv("GCS_AUTH_FILE" = "<access key>", "GCS_DEFAULT_BUCKET" = "bucket name")
+  if (nzchar(Sys.getenv("GCS_AUTH_FILE")) && nzchar(Sys.getenv("GCS_DEFAULT_BUCKET"))) {
+    return(invisible(TRUE))
+  }
+
+  testthat::skip("No GCS Credentials")
+}
+
 skip_without_aws_credentials <- function() {
   # -# Sys.setenv("AWS_ACCESS_KEY_ID" = "<access key>", "AWS_SECRET_ACCESS_KEY" = "<access secret>")
   if (nzchar(Sys.getenv("AWS_ACCESS_KEY_ID")) && nzchar(Sys.getenv("AWS_SECRET_ACCESS_KEY"))) {