+++ NEWS.md
@@ -1,6 +1,7 @@
 # Version 1.1.0.9000
 * Name clashes between function arguments and variables defined when memoising
   no longer occur (#43, @egnha).
+* Add Google Cloud Storage support via `cache_gcs()`
 
 # Version 1.1.0
 * Caches now hash the function body along with the arguments, to ensure